requirements.txt
fastapi
uvicorn[standard]
twilio
python-dotenv
httpx
python-multipart
pydantic
apscheduler
jinja2
